+++ libs/cardano-data/CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 1.2.2.0
 
 * Add `assocList`
+* Add module Data.MonoTuple
 
 ## 1.2.1.0